lib: use 64 bit sizes and offsets

+++ include/private/tree.h
@@ -21,10 +21,10 @@ int bp__tree_write_head(bp__writer_t* w, void* data);
 
 
 struct bp__tree_head_s {
-  uint32_t offset;
-  uint32_t config;
-  uint32_t page_size;
-  uint32_t hash;
+  uint64_t offset;
+  uint64_t config;
+  uint64_t page_size;
+  uint64_t hash;
 };
 
 #ifdef __cplusplus
